Add support for local parsing of YubiOTPs
This makes the OTP module a struct representing a YubiOTP. The
OTP can be parsed to hydrate fields on the struct. This permits
servers for validating and rejecting self-programmed tokens to be
written in Elixir.

The exisitng functions on the module have been refactored to make
use of the parsing functionality.

All new functionality has documentation and tests (in the form of
doctests).